Options menu Turning some code into basic functions for clarity Moving menu code to a new file
